refactored and changed hierarchy
