+++ src/views/sign/ForgetPw.js
@@ -7,13 +7,14 @@ import {
   Divider,
   TextField,
 } from '@material-ui/core'
-import { signStyle } from './signStyle'
-import SignLayout from "./SignLayout"
+import { signStyle } from './styles/signStyle'
+import SignLayout from "./layouts/SignLayout"
 
 import firebase from "firebase/app"
 import "firebase/auth";
 
-import { SignUpLink, SignInLink, ResendMailLink } from '../layouts/link-texts/SignLinks'
+import SignLink from './parts/SignLinks'
+import PATHS from '../../const/paths'
 
 
 const ForgetPw = () => {
@@ -69,11 +70,11 @@ const ForgetPw = () => {
       </Box>
       
       <Box textAlign={"right"} mb={4}>
-        <SignUpLink component="p" underline="always" />
+        <SignLink path={PATHS.SIGN_UP} />
         <br/>
-        <SignInLink component="p" underline="always" />
+        <SignLink path={PATHS.SIGN_IN} />
         <br/>
-        <ResendMailLink component="p" underline="always" />
+        <SignLink path={PATHS.RESEND_EMAIL} />
       </Box>
 
     </SignLayout>
